requirements.txt
flask==3.0.0
paddlepaddle==2.6.0
paddleocr==2.7.3
Pillow==10.1.0
gunicorn==21.2.0
numpy==1.24.3
